refactor: badrequest 패키지 생성
