+++ backend/src/main/java/com/cruru/applicant/exception/badrequest/ApplicantRejectException.java
@@ -1,4 +1,4 @@
-package com.cruru.applicant.exception;
+package com.cruru.applicant.exception.badrequest;
 
 import com.cruru.advice.badrequest.BadRequestException;
