+++ backend/src/main/java/com/cruru/applicant/service/ApplicantService.java
@@ -10,7 +10,7 @@ import com.cruru.applicant.controller.dto.QnaResponse;
 import com.cruru.applicant.domain.Applicant;
 import com.cruru.applicant.domain.repository.ApplicantRepository;
 import com.cruru.applicant.exception.ApplicantNotFoundException;
-import com.cruru.applicant.exception.ApplicantRejectException;
+import com.cruru.applicant.exception.badrequest.ApplicantRejectException;
 import com.cruru.process.controller.dto.ProcessSimpleResponse;
 import com.cruru.process.domain.Process;
 import com.cruru.process.domain.repository.ProcessRepository;
