+++ backend/src/main/java/com/cruru/choice/service/ChoiceService.java
@@ -3,8 +3,8 @@ package com.cruru.choice.service;
 import com.cruru.choice.controller.dto.ChoiceCreateRequest;
 import com.cruru.choice.domain.Choice;
 import com.cruru.choice.domain.repository.ChoiceRepository;
-import com.cruru.choice.exception.ChoiceEmptyBadRequestException;
-import com.cruru.choice.exception.ChoiceIllegalSaveException;
+import com.cruru.choice.exception.badrequest.ChoiceEmptyBadRequestException;
+import com.cruru.choice.exception.badrequest.ChoiceIllegalSaveException;
 import com.cruru.question.domain.Question;
 import com.cruru.question.domain.repository.QuestionRepository;
 import com.cruru.question.exception.QuestionNotFoundException;
