+++ backend/src/main/java/com/cruru/evaluation/domain/Evaluation.java
@@ -2,7 +2,7 @@ package com.cruru.evaluation.domain;
 
 import com.cruru.BaseEntity;
 import com.cruru.applicant.domain.Applicant;
-import com.cruru.evaluation.exception.EvaluationScoreException;
+import com.cruru.evaluation.exception.badrequest.EvaluationScoreException;
 import com.cruru.process.domain.Process;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
