+++ backend/src/main/java/com/cruru/process/exception/badrequest/ProcessCountException.java
@@ -1,4 +1,4 @@
-package com.cruru.process.exception;
+package com.cruru.process.exception.badrequest;
 
 import com.cruru.advice.badrequest.BadRequestException;
