+++ backend/src/main/java/com/cruru/process/service/ProcessService.java
@@ -13,10 +13,10 @@ import com.cruru.process.controller.dto.ProcessUpdateRequest;
 import com.cruru.process.controller.dto.ProcessesResponse;
 import com.cruru.process.domain.Process;
 import com.cruru.process.domain.repository.ProcessRepository;
-import com.cruru.process.exception.ProcessCountException;
-import com.cruru.process.exception.ProcessDeleteEndsException;
-import com.cruru.process.exception.ProcessDeleteRemainingApplicantException;
 import com.cruru.process.exception.ProcessNotFoundException;
+import com.cruru.process.exception.badrequest.ProcessCountException;
+import com.cruru.process.exception.badrequest.ProcessDeleteEndsException;
+import com.cruru.process.exception.badrequest.ProcessDeleteRemainingApplicantException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
