+++ backend/src/test/java/com/cruru/choice/service/ChoiceServiceTest.java
@@ -8,8 +8,8 @@ import com.cruru.applyform.domain.ApplyForm;
 import com.cruru.applyform.domain.repository.ApplyFormRepository;
 import com.cruru.choice.controller.dto.ChoiceCreateRequest;
 import com.cruru.choice.domain.Choice;
-import com.cruru.choice.exception.ChoiceEmptyBadRequestException;
-import com.cruru.choice.exception.ChoiceIllegalSaveException;
+import com.cruru.choice.exception.badrequest.ChoiceEmptyBadRequestException;
+import com.cruru.choice.exception.badrequest.ChoiceIllegalSaveException;
 import com.cruru.dashboard.domain.Dashboard;
 import com.cruru.dashboard.domain.repository.DashboardRepository;
 import com.cruru.question.domain.Question;
