+++ backend/src/test/java/com/cruru/evaluation/domain/EvaluationTest.java
@@ -2,7 +2,7 @@ package com.cruru.evaluation.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.cruru.evaluation.exception.EvaluationScoreException;
+import com.cruru.evaluation.exception.badrequest.EvaluationScoreException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
