+++ backend/src/test/java/com/cruru/process/service/ProcessServiceTest.java
@@ -22,9 +22,9 @@ import com.cruru.process.controller.dto.ProcessUpdateRequest;
 import com.cruru.process.controller.dto.ProcessesResponse;
 import com.cruru.process.domain.Process;
 import com.cruru.process.domain.repository.ProcessRepository;
-import com.cruru.process.exception.ProcessCountException;
-import com.cruru.process.exception.ProcessDeleteEndsException;
-import com.cruru.process.exception.ProcessDeleteRemainingApplicantException;
+import com.cruru.process.exception.badrequest.ProcessCountException;
+import com.cruru.process.exception.badrequest.ProcessDeleteEndsException;
+import com.cruru.process.exception.badrequest.ProcessDeleteRemainingApplicantException;
 import com.cruru.util.ServiceTest;
 import java.util.Comparator;
 import java.util.List;
